Split Api and Server into separate modules
